requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==20.3.0
bcrypt==3.2.0
blessed==1.17.11
boto==2.49.0
boto3==1.17.112
botocore==1.20.112
cached-property==1.5.2
cement==2.8.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
cryptography==3.1.1
debug-tools==2.6.20
debugtools==0.2.0
distlib==0.3.1
Django==3.1
django-admin-searchable-dropdown==1.2
django-cors-headers==3.7.0
django-crontab==0.7.1
django-debugtools==1.8
django-filter==2.4.0
django-rest-authtoken==2.1.3
djangorestframework==3.11.1
djangorestframework-api-key==2.0.0
docker==4.3.1
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
filelock==3.0.12
future==0.16.0
gunicorn==20.1.0
idna==2.10
importlib-metadata==2.1.3
iniconfig==1.1.1
jmespath==0.10.0
jsonschema==3.2.0
Markdown==3.3.4
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==1.4.6
packaging==20.9
paramiko==2.7.2
pathspec==0.5.9
Pillow==8.2.0
pipenv==2020.8.13
pluggy==0.13.1
py==1.10.0
pycparser==2.20
PyNaCl==1.4.0
pymongo==4.4.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.4
pytest-django==4.4.0
pytest-order==1.0.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-mock==1.9.3
s3transfer==0.4.2
semantic-version==2.5.0
six==1.15.0
sqlparse==0.3.1
sshtunnel==0.1.5
termcolor==1.1.0
texttable==1.6.3
toml==0.10.2
urllib3==1.25.11
virtualenv==20.1.0
virtualenv-clone==0.5.4
wcwidth==0.1.9
websocket-client==0.57.0
xmltodict==0.12.0
zipp==3.9.0
